requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.10
lxml==4.5.1
pytz==2020.1
requests==2.24.0
psycopg2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
